requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
--index-url https://mirrors.bfsu.edu.cn/pypi/web/simple

attrs==22.2.0
    # via pytest
beancount==2.3.5
    # via -r requirements.in
beautifulsoup4==4.11.1
    # via beancount
bottle==0.12.23
    # via beancount
cachetools==5.3.0
    # via google-auth
certifi==2022.12.7
    # via requests
chardet==5.1.0
    # via beancount
charset-normalizer==3.0.1
    # via requests
exceptiongroup==1.1.0
    # via pytest
google-api-core==2.11.0
    # via google-api-python-client
google-api-python-client==2.74.0
    # via beancount
google-auth==2.16.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
google-auth-httplib2==0.1.0
    # via google-api-python-client
googleapis-common-protos==1.58.0
    # via google-api-core
httplib2==0.21.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.4
    # via requests
iniconfig==2.0.0
    # via pytest
lxml==4.9.2
    # via beancount
packaging==23.0
    # via pytest
pluggy==1.0.0
    # via pytest
ply==3.11
    # via beancount
protobuf==4.21.12
    # via
    #   google-api-core
    #   googleapis-common-protos
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pymupdf==1.21.1
    # via -r requirements.in
pyparsing==3.0.9
    # via httplib2
pytest==7.2.1
    # via beancount
python-dateutil==2.8.2
    # via beancount
python-magic==0.4.27
    # via beancount
requests==2.28.2
    # via
    #   beancount
    #   google-api-core
rsa==4.9
    # via google-auth
six==1.16.0
    # via
    #   google-auth
    #   google-auth-httplib2
    #   python-dateutil
soupsieve==2.3.2.post1
    # via beautifulsoup4
tomli==2.0.1
    # via pytest
uritemplate==4.1.1
    # via google-api-python-client
urllib3==1.26.14
    # via requests
